requirements.txt
matplotlib==3.7.1
xlrd<2.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
streamlit==1.20.0
